requirements.txt
./vinted-api-wrapper
discord.py
databases[aiosqlite]
